refactor(ui): split editor timeline dock panel
